requirements.txt
beautifulsoup4==4.9.0
celery==5.0.0
cmake==3.17.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.6
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-debug-toolbar==2.1
django-email-verification==0.0.7
django-embed-video==1.3.3
django-excel-response2==2.0.8
django-heroku==0.3.1
django-js-asset==1.2.2
django-materializecss-form==1.1.17
django-mathfilters==1.0.0
oauthlib==3.1.0
Pillow==7.1.2
PyGetWindow==0.0.8
Pygments==2.6.1
PyTweening==1.0.3
requests==2.23.0
requests-oauthlib==1.3.0
Send2Trash==1.5.0
six==1.14.0
smtp==0.1.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
static-ranges==0.2.0
sqlparse==0.3.0
stripe==2.50.0
urllib3==1.25.9

